# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NPTC Catalogue Maintenance Platform",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/catalogue/entries/{business_key}/properties": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "summary": "An entry's recorded property values",
        "operationId": "read_properties_api_v1_catalogue_entries__business_key__properties_get",
        "parameters": [
          {
            "name": "business_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^NPTC-([0-9]{6,})$",
              "description": "The entry's public identifier, e.g. `NPTC-000247` (FR-03).",
              "examples": [
                "NPTC-000247"
              ],
              "title": "Business Key"
            },
            "description": "The entry's public identifier, e.g. `NPTC-000247` (FR-03)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyList"
                }
              }
            }
          },
          "401": {
            "description": "A credential was presented and could not be verified. Sending no credential at all is not an error on these endpoints.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No entry in the published catalogue has this business key. An entry that exists but is not published (draft, deprecated or withdrawn) is reported identically, on purpose - a distinguishable response would confirm the key exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A query or path parameter was unprocessable - a business key that is not `NPTC-nnnnnn`, a blank search query, a cursor this API did not issue (including one issued for a different `q` or a different filter set), a `limit` outside its range, or a `filter.*` parameter naming a facet this endpoint does not offer, an operator the facet does not support, or a value the property cannot hold. A filter is never silently ignored.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/catalogue/entries/{business_key}/properties/{key}": {
      "put": {
        "tags": [
          "catalogue-admin"
        ],
        "summary": "Replace a property's recorded values on a catalogue entry",
        "operationId": "save_property_api_v1_catalogue_entries__business_key__properties__key__put",
        "parameters": [
          {
            "name": "business_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^NPTC-([0-9]{6,})$",
              "description": "The entry's public identifier, e.g. `NPTC-000247` (FR-03).",
              "examples": [
                "NPTC-000247"
              ],
              "title": "Business Key"
            },
            "description": "The entry's public identifier, e.g. `NPTC-000247` (FR-03)."
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavePropertyValuesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyValuesWriteResult"
                }
              }
            }
          },
          "401": {
            "description": "No credential, or one that could not be verified.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The caller is authenticated but does not hold `catalogue.edit_published`, or holds it but has not completed the MFA step-up this permission requires (the response then also carries a `WWW-Authenticate` step-up challenge).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No catalogue entry, or no property definition, matches the given identifier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The submitted `expected_row_version` no longer matches the entry's current `row_version` (FR-38) - someone else changed this entry since it was loaded. Carries `business_key`, `expected_row_version`, `current_row_version`, `conflicts[]` (each with `field`, `submitted` and `current`) and `changed_by`/`changed_at`, so the caller can reconcile rather than retry blind.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/VersionConflictResponse"
                    }
                  ],
                  "title": "Response 409 Save Property Api V1 Catalogue Entries  Business Key  Properties  Key  Put"
                }
              }
            }
          },
          "422": {
            "description": "The `reason` is missing or low-information (FR-37), the write targets a deprecated property (FR-11), or one or more submitted values fail their property's JSON Schema, cardinality bound, or FR-89's specimen cross-field check - `issues[]` then names the `property_key`, `label` and `ordinal` of each failing value.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/PropertyValidationResponse"
                    }
                  ],
                  "title": "Response 422 Save Property Api V1 Catalogue Entries  Business Key  Properties  Key  Put"
                }
              }
            }
          },
          "500": {
            "description": "The definition's own stored `datatype` no longer matches a registered handler - a data integrity fault in the definition, not a caller mistake. Not produced by anything a well-formed request can trigger on its own.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalogue/entries/bulk/properties/{key}": {
      "post": {
        "tags": [
          "catalogue-admin"
        ],
        "summary": "Replace a property's recorded values across many catalogue entries",
        "operationId": "save_property_bulk_api_v1_catalogue_entries_bulk_properties__key__post",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkSavePropertyValuesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkSavePropertyValuesResult"
                }
              }
            }
          },
          "401": {
            "description": "No credential, or one that could not be verified.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The caller is authenticated but does not hold `catalogue.edit_published`, or holds it but has not completed the MFA step-up this permission requires (the response then also carries a `WWW-Authenticate` step-up challenge).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No property definition matches `key`. An unknown `business_key` among `entries` is a per-entry `not-found` outcome in the 200 response, never a 404 for the whole request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The `reason` is missing or low-information (FR-37), the write targets a deprecated property (FR-11), or one or more submitted values fail their property's JSON Schema, cardinality bound, or FR-89's specimen cross-field check - `issues[]` then names the `property_key`, `label` and `ordinal` of each failing value.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/HTTPValidationError"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/PropertyValidationResponse"
                    }
                  ],
                  "title": "Response 422 Save Property Bulk Api V1 Catalogue Entries Bulk Properties  Key  Post"
                }
              }
            }
          },
          "500": {
            "description": "The definition's own stored `datatype` no longer matches a registered handler - a data integrity fault in the definition, not a caller mistake. Not produced by anything a well-formed request can trigger on its own.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkSavePropertyValuesRequest": {
        "properties": {
          "values": {
            "items": {
              "$ref": "#/components/schemas/PropertyValueItemRequest"
            },
            "type": "array",
            "title": "Values"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/BulkPropertyEntryTarget"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "values",
          "reason",
          "entries"
        ],
        "title": "BulkSavePropertyValuesRequest",
        "description": "The body of `POST /catalogue/entries/bulk/properties/{key}`\n(FR-39). `values` is the one set every named entry ends up\nholding - a whole-set replace, identical to the singular route's own\nsemantics, applied across `entries` rather than one."
      },
      "BulkSavePropertyValuesResult": {
        "properties": {
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/BulkPropertyOutcomeItem"
            },
            "type": "array",
            "title": "Outcomes"
          },
          "applied": {
            "type": "integer",
            "title": "Applied"
          },
          "unchanged": {
            "type": "integer",
            "title": "Unchanged"
          },
          "conflict": {
            "type": "integer",
            "title": "Conflict"
          },
          "not_found": {
            "type": "integer",
            "title": "Not Found"
          }
        },
        "type": "object",
        "required": [
          "outcomes",
          "applied",
          "unchanged",
          "conflict",
          "not_found"
        ],
        "title": "BulkSavePropertyValuesResult",
        "description": "The per-entry outcome list, plus its own tallies. Always a 200: the\nrequest was authorised, well-formed, and fully processed, and the\noutcomes *are* the representation - including a batch where every\nentry conflicted (a whole-request 409 would have to discard the applied\nentries' new `row_version`s, the one thing a retrying client needs)."
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Every refusal this API makes has this shape - one sentence saying\nwhat to do next, and deliberately nothing else. It never names a role,\na permission or an internal identifier (FR-44, NFR-04)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PropertyList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PropertyValue"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PropertyList"
      },
...
      "PropertyValidationResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/PropertyIssueItem"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "issues"
        ],
        "title": "PropertyValidationResponse",
        "description": "FR-09/FR-10/FR-88/FR-89's 422 body: `PropertyValidationError`'s\n`issues[]`, declared as a model rather than a hand-built dict - a\nrouter naming this in its `responses=` puts the real `issues[]` shape\nin `docs/api/openapi.json`, matching `VersionConflictResponse`/\n`DesignationCollisionResponse`'s own precedent, so a generated client\ntypes the field-level detail instead of a bare `{detail}`."
      },
...
      "PropertyValuesWriteResult": {
        "properties": {
          "values": {
            "items": {
              "$ref": "#/components/schemas/PropertyValue"
            },
            "type": "array",
            "title": "Values"
          },
          "row_version": {
            "type": "integer",
            "title": "Row Version"
          }
        },
        "type": "object",
        "required": [
          "values",
          "row_version"
        ],
        "title": "PropertyValuesWriteResult",
        "description": "The property's values after the write, plus the entry's new\n`row_version` - mirroring `catalogue_designations.\nAmendDesignationResult`, so an editing client never has to re-fetch the\nentry just to learn its next lock token."
      },
...
      "SavePropertyValuesRequest": {
        "properties": {
          "values": {
            "items": {
              "$ref": "#/components/schemas/PropertyValueItemRequest"
            },
            "type": "array",
            "title": "Values"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "expected_row_version": {
            "type": "integer",
            "title": "Expected Row Version"
          }
        },
        "type": "object",
        "required": [
          "values",
          "reason",
          "expected_row_version"
        ],
        "title": "SavePropertyValuesRequest",
        "description": "The body of `PUT /catalogue/entries/{business_key}/properties/{key}`.\n\n`values` is the complete, intended value set for this property after\nthe write - `save_property_values` replaces the whole set, never a\ndiff (see that function's own module docstring)."
      },
...
      "VersionConflictResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "business_key": {
            "type": "string",
            "title": "Business Key"
          },
          "expected_row_version": {
            "type": "integer",
            "title": "Expected Row Version"
          },
          "current_row_version": {
            "type": "integer",
            "title": "Current Row Version"
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/FieldConflictItem"
            },
            "type": "array",
            "title": "Conflicts"
          },
          "changed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed By"
          },
          "changed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed At"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "business_key",
          "expected_row_version",
          "current_row_version",
          "conflicts",
          "changed_by",
          "changed_at"
        ],
        "title": "VersionConflictResponse",
        "description": "FR-38's 409 body: a stale `expected_row_version` on an entry-level\nwrite.\n\nFR-38's rationale rejects silent last-write-wins \"because it produces an\naudit trail that records a change that was immediately and invisibly\ndiscarded\", so the refusal has to let the caller *reconcile* rather than\nretry blind. `conflicts` is empty where the caller's submitted values do\nnot themselves overlap what moved - still a refusal, because the version\nis the contract regardless - which is why `current_row_version` and\n`changed_by`/`changed_at` are populated even then.\n\n`changed_by` is a display name, never the actor's internal id\n(NFR-04/NFR-26), and is `null` for a system-initiated change or an\naccount since pseudonymised on closure (NFR-17)."
      }
    }
  }
}